Update exports of postcss files
By not having a default export when there is no css-modules class name
mapping we avoid several "var undefined$1 = undefined;" lines when using
modern rollup with no-treeshake.

When there was a file with no css-modules class mapping this previously
resulted in `export default undefined` which doesn't make any sense. Now
we just omit the default export. Other behaviour remains the same as
current.

Note that we do subtly different things when extracting vs injecting -
when extracting we never export the plain css in non-modules land. We
could make that consistent in the future but I didn't want to address
that now.